Restructure files.
